ExUnit.Diff: Escape pinned values when computing diff
Take these two tests:

```elixir
  test "correctly colored" do
    assert [{:foo}] = [{:bar}]
  end

  test "incorrectly colored" do
    val = [{:foo}]
    assert ^val = [{:bar}]
  end
```

When diffing the match without a pin, `diff_value/3` would be passed the
quoted version of `[{:foo}]` (`[{:{}, [], [:foo]}]`) as the left argument,
allowing it to recurse deeper into the structure to find the correct diff.

When diffing the match with the pin, `diff_value/3` would be passed `[{:foo}]`
instead, prohibiting it from recursively refining the diff, leading to overly
broad diffs. The fix is to explicitly escape the pin value in the call to
`diff_value/3` inside of `diff_pin/3`.

This fixes #13348