Translate `:undefined` URI port to `nil`
Resolves #13462

When a url string is schema less and the host is followed by a colon
without setting the actual port, `:uri_string.parse/1` returns the
port to be `:undefined`. As long as the url string is parseable,
we should translate `:undefiend` to `nil`, in order to ensure we return a valid URI struct.