Add some signposting to the String docs for pattern matching
As a user of the docs, when looking for information on how to pattern
match with strings and binaries the first place one looks is the String
docs.

Currently there isn't anything here for pattern matching besides
a small example matching a character. The user would either have to
go from the reference section to the patterns and guards guide
or go to the <<>>/1 special forms docs.

This change just adds a link to the special forms docs and the guide
to signpost the user to those resources.