Make clear the guide also talks about use
Browsing the docs I found the page and found it weird that it
didn't mention use (as the 4 of them go together), reading the
page I realized that it _does_ mention `use` but it's just not
mentioned in the headline so this seemed like an easy fix.

I did _not_ modify the file name, although I think it'd be better
to also rename the file (and references to it). However, I was
worried about breaking outside links to https://hexdocs.pm/elixir/alias-require-and-import.html
I'm looking for guidance on:
* is the breakage ok
* should we just not break
* should we break and maybe leave a note in the existing file
  that documenation moved (and hide from the getting started)