Fix issues with escaping and arguments parsing
PowerShell have some weird rules for parsing arguments. We need to take
care of strings that have spaces, because the space char is considered a
separator for arguments. We need to surround these strings with
double-quotes, but we also need to escape the quotes inside first.

When it's a normal execution, this escaping is just "doubling" the
double-quotes. But when the execution is for the `--pipe-to` argument,
this needs to be an Unix escaping done twice for double-quotes.