Add releases for main & maintenance branches (#13829)
* Add releases for main & maintenance branches

Before this patch, release.yml workflow was only compiling and
publishing builds on tag pushes.

This patch additionally handles branch pushes (main, v1.17, etc) and
publishes them to corresponding releases (main-latest, v1.17-latest,
etc).

When we first looked at using GitHub releases for main builds, we
discarded the idea because we didn't want to keep re-creating (and thus
sending out notifications) releases. There was no API to update the SHA
the release was pointing too either. It's obvious in hindsight but all
we need to do is to update the _tag_ with the latest SHA and force-push
the tag that the release is pointing to.

We now have two places where we build main and maintenance branches,
release.yml and builds.hex.pm.yml, so in future commits we should look
into unifying the two.

* Update .github/workflows/release.yml