Add tests for parsing fractional durations between -1 and +1 seconds
Extend tests for existing fractional durations parsing behaviour.

Parsing "PT0,6S" should result in a positive microsecond, which will be
fixed in the next commit.

The parser currently accepts both a negative prefix "-PT" and a negative
value in the components "-0,6S". Combining both into "-PT-0,6S" is also
accepted by the parser, resulting in an overall positive duration.
Although the practical value of double negative durations is dubious,
the accepting nature of the parser is tested.

The JavaScript libraries that I tested disagree whether "-PT-0,6S"
represents a positive or negative duration, "moment" and "tinyduration"
reported positive, "luxon" (the successor to moment) reported negative.
If the parser accepts double negatives at all, resulting in anything but
a positive duration would be unintuitive.