Provide some more pointers for logger overload protection
The previous docs didn't mention that logger may drop messages
entirely, which I think is important to call out (and not just
because I finished a long debugging journey to find these :) ).

This should give a fair overview of the built-in overload
protections, similar to what I tried to do in erlang/otp#8855

I think this can really help folks be aware of some of these.