Clarify the interplay of `Stream` and `Enum`
This is based on trying to clarify/make more obvious a
missunderstanding I held until ~yesterday.

Namely I thought that calling an `Enum` function on a stream would
result in the equivalent of `Enum.to_list` and hence all stream
benefits would be forfeit. Yes, it should be clear seeing
`Enum.take` work that this isn't the case, but it's not always
so easy.

This turned out wrong as @benwilson512 pointed out eloquently here:
https://elixirforum.com/t/why-is-stream-reduce-while-missing/34422/2

Not sure if this is the best place or way to highlight this in the
docs, but I thought I'd give it a shot. One could argue that this
documentation belongs in the `Enum` module, but for `Stream`
documentation that module is the primary source - plus it's
more about `Stream`s implementation of `Enumerable`.