Make protocol errors pretty
A common complaint I here from people is that error messages in Elixir
are not pretty. Generally, this seems to be only the case when there are
large structs in codebases that render in one line and are hard to read.

This PR changes the behavior to pretty inspect the structs in the case
that a protocol is not implemented for some value.

The downside here is that the logging will no longer be in one line, but
I think this is a fair tradeoff for legibility