Add override_for to dep options
override_for gives more information about when deps should explicitly be overridden or not.

If a dep no longer needs an override, a message will be written indicating so.
Alternatively, if a dep needs an override when a dep is already using override_for, it will alert the dev.

With override: true set, if a user were to add a new dep that depends on the overridden app, there would be no indication of a potential issue.