Add :single_column option to IEx.Helpers.exports (second edit).
After previous commit, the code was working fine, but it was reading backwards. 

Original description:
Adds a `:single_column` option to `IEx.Helpers.exports/2` to print exported functions/macros in a single column. This improves readability for modules with long function names or when working in terminals with limited horizontal space.