Initial attempt at embeddable regex objects
This patch, which works on 1.18.x but does not work on 1.19
is an attempt to implement String.Chars protocol and also
a Regex.to_string() and Regex.modifiers() and Regex.to_string!()
and Regex.modifiers!() functions.

The idea is to make it possible to safely embed precompiled
regexes into other regexes in a similar way as that supported
by perl.  The general idea is that ~r/foo/x turns into
"(?x-ims:foo\n)", and etc. Thus it should match the same as
it would have in its original form when it is embedded
into a pattern which has a different set of modifiers.

For review by Jose.