regex.ex - add support for to_embed()
to_embed(regex,strict) returns an embeddable representation of regex.
For instance ~r/foo/i can be represented as ~r/(?i-msx:foo)/.

If the option :strict is true (the default) then it will throw an
ArgumentError if the regex was compiled with an option/modifier which
cannot be represented as an embeddable pattern.  If :strict is false
then any unembeddable options will be silently ignored.  This may be
perfectly reasonable, for intance the wrapped pattern may be compiled
with the same modifiers as the pattern, or reusing the pattern without
the unembeddable modifiers may not change its semantics.