File.LinkError - swap order of new and existing in error message
The api for creating links is ln(existing, new) but the old error text
said "from existing to new", where I believe it should say "from new to
existing".

Likely this error came from the example using "source.txt" and
"target.txt" and it being natural to say "to target", so this patch also
changes the example to use "existing.txt" and "link.txt" as the file
names so it reads more naturally.