Improve map_line_empty and emptiness algoriths
The key change is that map_line_empty was checking the whole tree no matter what, due to an or condition that should not be checked as the previous Enum.all? are enough to conclude.

Also clarified the invariant that allows not to check the positives fields as a base case (still in map_line_empty?)