Clarify Scope of Open Source Policy
Update OPEN_SOURCE_POLICY.md to clarify that the policy applies to all
projects under the Elixir organization, and not for any  project written in
the Elixir programming language.

This change helps to avoid ambiguity around what contributions are covered
by the policy and ensures that contributors understand which guidelines
apply to their work.