Fix instructions for  building docs from release sources
The following instructions do not work because the source code version
of the release create a folder with the version in it.

```console
❯ unzip -l v1.19.3.zip
Archive:  v1.19.3.zip
427d21e410d1846c232d0a0b0e87cce629884dfa
  Length      Date    Time    Name
---------  ---------- -----   ----
        0  11-13-2025 11:33   elixir-1.19.3/
      502  11-13-2025 11:33   elixir-1.19.3/.formatter.exs
...
```

So the instructions do not work because it expects a folder named `elixir/`.

Instructions have been updated to work from the Git repo as well as from the
downloaded release source files.