Propagate BDD leaf nodes down the tree on intersections
LazyBDDs allowed us to represent all operations lazily
but the issue is that we want to propagate intersections
eagerly, as they allow us to cut down the number of nodes
in the tree. So whenever we have a leaf node, we propagate
it throughout the whole tree, cutting down the number of
nodes considerably and optimizing type checking.

Before:

    [profile] Finished group pass check of 1765 modules in 13506ms

After:

    [profile] Finished group pass check of 1765 modules in 2454ms